requirements.txt
pyzotero
python-slugify
